Add default value to skip_package_upload on Manifest.retrieve
Previously, the new variable did not have a default value associated
with it, and the PR which added the option did not update all the calls
to Manifest.retrieve to include the new variable.

Instead of updating all of the references to pass in false, have it
simply use a default value of false.

This was an issue for a number of the commands other than "upload".