Fix --fail-if-exists
When --fail-if-exists is in effect, the code makes a HEAD request to
check if the object exists and has the same md5, but uses the results of
that request to write the name of the object via the `public_url` method
(that only applies to GET request results).

As we already have the filename in the method, just use that directly in
the message instead of an object property.

Fixes #141.

Signed-off-by: L. Alberto Giménez <alberto.gimenez@capside.com>