Clarify --fail-if-exists effect
When the switch is enabled, and a package is uploaded with the same
version **and contents**, it will not actually fail.

It only has effect when the uploaded package matches an existing package
in the pool or manifest, but the contents are different.

Clarify commandline reference as well as README.

Closes #136.

Signed-off-by: L. Alberto Giménez <alberto.gimenez@capside.com>